+++ src/components/DataExplorerVisualization.jsx
