+++ src/components/layout/nav/InboxDropdownMenu.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Radium from 'radium';
 
-import NavDropdownMenu from './nav-dropdown-menu';
+import NavDropdownMenu from './NavDropdownMenu';
 
 class InboxDropdownMenu extends React.Component {
   render() {
