+++ src/components/layout/sidebar/Menu.jsx
@@ -2,7 +2,7 @@ import _ from 'lodash';
 import React from 'react';
 import Radium from 'radium';
 
-import ModuleItem from './menu-item';
+import ModuleItem from './ModuleItem';
 
 @Radium
 class Menu extends React.Component {
