+++ src/index.js
@@ -9,12 +9,11 @@ import { Provider } from 'react-redux';
 import { DevTools, DebugPanel, LogMonitor } from 'redux-devtools/lib/react';
 
 import configureStore from './store';
-import Settings from './containers/settings';
 
 import Dashboard from './containers/Dashboard';
-import UserManager from './containers/user-manager';
-import Login from './containers/login';
-import DataExplorer from './containers/data-explorer';
+import UserManager from './containers/UserManager';
+import Login from './containers/Login';
+import DataExplorer from './containers/DataExplorer';
 
 const store = configureStore();
 
@@ -30,7 +29,6 @@ class Root extends React.Component {
         <Provider store={store}>
           <Router history={browserHistory}>
             <Route path="/" component={Dashboard}/>
-            <Route path="/settings" component={Settings}/>
             <Route path="login" component={Login}/>
             <Route path="user-manager/:filterId/:userId" component={UserManager}/>
             <Route path="user-manager/:filterId" component={UserManager}/>
