+++ webpack.config.dev.js
@@ -18,7 +18,7 @@ module.exports = {
   ],
   module: {
     loaders: [{
-      test: /\.js$/,
+      test: /\.jsx?$/,
       loader: 'babel',
       exclude: /node_modules/,
       include: path.join(__dirname, 'src')
